M4-M7: hotkey backends verified end-to-end, TTS engines (Kokoro/ort2 vendored port compiled; Piper+system fallbacks linked), Swift accel-helper sources (M6), packaging (sign/notarize script, AppImage/deb/Flatpak collateral, udev rule, linux-setup), docs (README privacy, LICENSES, QA checklist)
- full app binary builds AND runs on Linux (whisper.cpp + llama.cpp linked); `local-ai-flow --doctor` verified in-container
- all 45 unit tests green; macOS backend cross-checked on aarch64-apple-darwin
- ort prebuilt-binaries glibc>=2.38 note documented (older distros: load-dynamic or skip tts-kokoro)

+++ crates/laf-engines/src/tts_kokoro.rs
@@ -60,7 +60,7 @@ impl KokoroEngine {
             self.model_path.to_string_lossy().as_ref(),
             self.tokenizer_path.to_string_lossy().as_ref(),
         )
-        .with_sample_rate(KOKORO_RATE as i32)
+        .with_sample_rate(KOKORO_RATE)
         .with_max_tokens_length(512);
         let tts = KokoroTTS::with_config(config)
             .map_err(|e| EngineError::Tts(format!("failed to load Kokoro: {e}")))?;
